criado arquivo conftest para utilizar as fixtures em outros arquivos
close #27
