+++ .idea/workspace.xml
@@ -2,7 +2,11 @@
 <project version="4">
   <component name="ChangeListManager">
     <list default="true" id="117c5f2e-87da-4d9f-ad6d-d7a0634c23fd" name="Changes" comment="corrigido arquivo python-app.yml">
+      <change afterPath="$PROJECT_DIR$/libpythonpro/spam/main.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/libpythonpro/tests/test_spam/conftest.py" afterDir="false" />
+      <change afterPath="$PROJECT_DIR$/libpythonpro/tests/test_spam/test_envio_para_a_base_de_usuarios.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/libpythonpro/spam/db.py" beforeDir="false" afterPath="$PROJECT_DIR$/libpythonpro/spam/db.py" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/libpythonpro/tests/test_spam/test_usuarios.py" beforeDir="false" afterPath="$PROJECT_DIR$/libpythonpro/tests/test_spam/test_usuarios.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
@@ -33,33 +37,34 @@
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent">{
-  &quot;keyToString&quot;: {
-    &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
-    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
-    &quot;settings.editor.selected.configurable&quot;: &quot;com.jetbrains.python.configuration.PyIntegratedToolsModulesConfigurable&quot;
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "Git.Branch.Popup.ShowAllRemotes": "true",
+    "RunOnceActivity.OpenProjectViewOnStart": "true",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "settings.editor.selected.configurable": "com.jetbrains.python.configuration.PyIntegratedToolsModulesConfigurable"
   }
-}</component>
+}]]></component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
       <recent name="$PROJECT_DIR$/libpythonpro/tests" />
       <recent name="$PROJECT_DIR$/libpythonpro/tests/test_spam" />
     </key>
   </component>
-  <component name="RunManager" selected="Python tests.pytest in test_usuarios.py">
-    <configuration name="pytest for libpythonpro.tests.test_spam.test_enviador_de_email.test_remetente_invalido" type="tests" factoryName="py.test" temporary="true" nameIsGenerated="true">
+  <component name="RunManager" selected="Python tests.pytest for test_envio_para_a_base_de_usuarios.test_envio_de_spam">
+    <configuration name="pytest for test_envio_para_a_base_de_usuarios.test_envio_de_spam" type="tests" factoryName="py.test" temporary="true" nameIsGenerated="true">
       <module name="libpythonpro" />
       <option name="INTERPRETER_OPTIONS" value="" />
       <option name="PARENT_ENVS" value="true" />
       <option name="SDK_HOME" value="" />
-      <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/libpythonpro/tests" />
+      <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/libpythonpro/tests/test_spam" />
       <option name="IS_MODULE_SDK" value="true" />
       <option name="ADD_CONTENT_ROOTS" value="true" />
       <option name="ADD_SOURCE_ROOTS" value="true" />
       <option name="_new_keywords" value="&quot;&quot;" />
       <option name="_new_parameters" value="&quot;&quot;" />
       <option name="_new_additionalArguments" value="&quot;&quot;" />
-      <option name="_new_target" value="&quot;libpythonpro.tests.test_spam.test_enviador_de_email.test_remetente_invalido&quot;" />
+      <option name="_new_target" value="&quot;test_envio_para_a_base_de_usuarios.test_envio_de_spam&quot;" />
       <option name="_new_targetType" value="&quot;PYTHON&quot;" />
       <method v="2" />
     </configuration>
@@ -95,19 +100,19 @@
       <option name="_new_targetType" value="&quot;PYTHON&quot;" />
       <method v="2" />
     </configuration>
-    <configuration name="pytest in test_enviador_de_email.py" type="tests" factoryName="py.test" temporary="true" nameIsGenerated="true">
+    <configuration name="pytest in test_envio_para_a_base_de_usuarios.py" type="tests" factoryName="py.test" temporary="true" nameIsGenerated="true">
       <module name="libpythonpro" />
       <option name="INTERPRETER_OPTIONS" value="" />
       <option name="PARENT_ENVS" value="true" />
       <option name="SDK_HOME" value="" />
-      <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/libpythonpro/tests" />
+      <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/libpythonpro/tests/test_spam" />
       <option name="IS_MODULE_SDK" value="true" />
       <option name="ADD_CONTENT_ROOTS" value="true" />
       <option name="ADD_SOURCE_ROOTS" value="true" />
       <option name="_new_keywords" value="&quot;&quot;" />
       <option name="_new_parameters" value="&quot;&quot;" />
       <option name="_new_additionalArguments" value="&quot;&quot;" />
-      <option name="_new_target" value="&quot;$PROJECT_DIR$/libpythonpro/tests/test_spam/test_enviador_de_email.py&quot;" />
+      <option name="_new_target" value="&quot;$PROJECT_DIR$/libpythonpro/tests/test_spam/test_envio_para_a_base_de_usuarios.py&quot;" />
       <option name="_new_targetType" value="&quot;PATH&quot;" />
       <method v="2" />
     </configuration>
@@ -129,11 +134,11 @@
     </configuration>
     <recent_temporary>
       <list>
+        <item itemvalue="Python tests.pytest for test_envio_para_a_base_de_usuarios.test_envio_de_spam" />
+        <item itemvalue="Python tests.pytest in test_envio_para_a_base_de_usuarios.py" />
         <item itemvalue="Python tests.pytest in test_usuarios.py" />
         <item itemvalue="Python tests.pytest for test_usuarios.test_listar_usuario" />
         <item itemvalue="Python tests.pytest for test_usuarios.test_salvar_usuario" />
-        <item itemvalue="Python tests.pytest for libpythonpro.tests.test_spam.test_enviador_de_email.test_remetente_invalido" />
-        <item itemvalue="Python tests.pytest in test_enviador_de_email.py" />
       </list>
     </recent_temporary>
   </component>
@@ -293,7 +298,14 @@
       <option name="project" value="LOCAL" />
       <updated>1667767136399</updated>
     </task>
-    <option name="localTasksCounter" value="22" />
+    <task id="LOCAL-00022" summary="consertado escopos de fixtures&#10;&#10;close #25">
+      <created>1667769858858</created>
+      <option name="number" value="00022" />
+      <option name="presentableId" value="LOCAL-00022" />
+      <option name="project" value="LOCAL" />
+      <updated>1667769858858</updated>
+    </task>
+    <option name="localTasksCounter" value="23" />
     <servers />
   </component>
   <component name="Vcs.Log.Tabs.Properties">
@@ -323,6 +335,7 @@
     <MESSAGE value="criado teste enviador de email&#10;&#10;close #21" />
     <MESSAGE value="emulado teste com banco de dados&#10;&#10;close #23" />
     <MESSAGE value="emulado teste com banco de dados, implementado conexão, sessão e usuário&#10;&#10;close #23" />
-    <option name="LAST_COMMIT_MESSAGE" value="emulado teste com banco de dados, implementado conexão, sessão e usuário&#10;&#10;close #23" />
+    <MESSAGE value="consertado escopos de fixtures&#10;&#10;close #25" />
+    <option name="LAST_COMMIT_MESSAGE" value="consertado escopos de fixtures&#10;&#10;close #25" />
   </component>
 </project>
