+++ libpythonpro/spam/main.py
@@ -1,0 +1,10 @@
+class Enviador_de_spam:
+    def __init__(self, sessao, enviador):
+        self.sessao = sessao
+        self.enviador = enviador
+
+    def enviar_emails(self, remetente, assunto, corpo):
+        pass
+
+
+
